requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.8
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-redis==4.10.0
django-rest-framework-mongoengine==3.3.1
django-reversion==3.0.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
future==0.17.1
httplib2==0.12.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mongoengine==0.16.3
mysql-connector==2.1.6
Pillow==5.4.1
pycryptodome==3.7.2
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.3
pytz==2018.9
raven==6.10.0
redis==3.0.1
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
XlsxWriter==1.1.2
xlwt==1.3.0
django-cors-headers==2.4.0